refactor: Reorganizando dependencias
