+++ src/main/java/challenge/domain/workflow/activity/usecase/simple/GetClientLoanRateActivity.java
@@ -1,4 +1,4 @@
-package challenge.domain.workflow.activity.core.simple;
+package challenge.domain.workflow.activity.usecase.simple;
 
 import challenge.domain.service.InterestRateService;
 import lombok.RequiredArgsConstructor;
